agregado de Listado Basico y cambio de double a Big decimal para mayor precision

+++ src/main/java/com/proyecto_promhil/proyecto_promhil/model/Compra.java
@@ -37,5 +37,5 @@ public class Compra {
 
     @NotNull
     @Column(name = "total", nullable = false, precision = 10, scale = 2)
-    private Double total;
+    private BigDecimal total;
 }
